expand database so we can't start making a weekly plan

+++ MealPlan.Logic.DomainObject/Models/MacroInfo.cs
@@ -2,7 +2,7 @@
 {
     public class MacroInfo
     {
-        public decimal Calories { get; set; } = 0;
+        public int Calories { get; set; } = 0;
         public decimal Carbs { get; set; } = 0;
         public decimal Fats { get; set; } = 0;
         public decimal Protein { get; set; } = 0;
